+++ WeatherApp/controller/ViewController.swift
@@ -259,25 +259,5 @@ class ViewController: UIViewController, UIScrollViewDelegate {
         print("\(location.latitude) | \(location.longitude)")
         weatherProvider.request(location: location)
     }
-    
 
-}
-
-
-extension Date {
-    
-    func dayOfWeek() -> String? {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "EEEE"
-        dateFormatter.locale = Locale(identifier: "en_EN")
-        return dateFormatter.string(from: self).lowercased()
-    }
-    
-    func dayOfMonth() -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "MMM dd"
-        dateFormatter.locale = Locale(identifier: "en_EN")
-        return dateFormatter.string(from: self).uppercased()
-    }
-    
 }
