+++ WeatherApp/model/WeatherJSONModel.swift
